requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0
torch>=1.9.0
torchvision>=0.10.0
svglib>=1.1.0
svgwrite>=1.4.0
matplotlib>=3.5.0
scipy>=1.7.0
networkx>=2.6.0
shapely>=1.8.0
potrace>=0.1.0
flask>=2.0.0
flask-login>=0.6.0
flask-wtf>=1.0.0
werkzeug>=2.0.0
python-dotenv>=0.19.0
vtracer>=0.6.0
bcrypt>=3.2.0
python-magic>=0.4.24
flask-limiter>=2.1.0

# Async Processing
celery>=5.3.0
redis>=4.5.0
flower>=2.0.0

# Testing dependencies
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0
pytest-benchmark>=4.0.0
coverage>=7.0.0
factory-boy>=3.2.0
fake-it>=0.3.0
responses>=0.21.0
selenium>=4.0.0
locust>=2.0.0
psutil>=5.9.0
memory-profiler>=0.60.0